BIGINT USER_ID

+++ app/models/users/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String
+from sqlalchemy import Column, Integer, String, BIGINT
 from config.database import Base
 
 
@@ -6,7 +6,7 @@ class UserModel(Base):
     __tablename__ = "users"
 
     id = Column(Integer, primary_key=True)
-    user_id = Column(Integer, unique=True)
+    user_id = Column(BIGINT, unique=True)
     first_name = Column(String)
     last_name = Column(String)
     username = Column(String)
